+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/ClaimPaymentAmount.java
@@ -5,7 +5,7 @@ import org.hl7.fhir.r4.model.ExplanationOfBenefit;
 
 public class ClaimPaymentAmount {
   @Column(name = "clm_pmt_amt")
-  private float claimPaymentAmount;
+  private double claimPaymentAmount;
 
   ExplanationOfBenefit.PaymentComponent toFhir() {
     return new ExplanationOfBenefit.PaymentComponent()
